requirements.txt
python-telegram-bot>=21.0
aiohttp>=3.9.0
sqlalchemy>=2.0.0
aiosqlite>=0.22.0
python-dotenv>=1.0.0